chore: linter fixes
